finance_dashboard_backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13
psycopg2-binary>=2.9
python-dotenv>=0.19
djangorestframework-simplejwt>=5.0